requirements.txt
uvicorn
asgiref
starlette
pdbfixer
shiny
pandas
numpy
requests
scipy
fair-esm
matplotlib
biopython
biotite
scikit-learn
optuna
seaborn
openpyxl
py3Dmol
gpytorch
